pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["leo-guinan <leo@copy.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.4"}
tweepy = "^4.9.0"
python-decouple = "^3.6"
django-cors-headers = "^3.12.0"
djangorestframework-simplejwt = "^5.1.0"
djangorestframework-api-key = "2.*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
